+++ dflib/src/main/java/com/nhl/dflib/accumulator/IntAccumulator.java
@@ -68,7 +68,11 @@ public class IntAccumulator implements Accumulator<Integer> {
     public void setInt(int pos, int value) {
 
         if (pos >= size) {
-            throw new IndexOutOfBoundsException(pos + " is out of bounds for " + size);
+            size = pos + 1;
+        }
+
+        if (size >= data.length) {
+            expand(Math.max(data.length * 2, size));
         }
 
         data[pos] = value;