+++ dflib/src/main/java/com/nhl/dflib/accumulator/ObjectAccumulator.java
@@ -48,7 +48,11 @@ public class ObjectAccumulator<T> implements Accumulator<T> {
     @Override
     public void set(int pos, T v) {
         if (pos >= size) {
-            throw new IndexOutOfBoundsException(pos + " is out of bounds for " + size);
+            size = pos + 1;
+        }
+
+        if (size >= data.length) {
+            expand(Math.max(data.length * 2, size));
         }
 
         data[pos] = v;